+++ packages/vexide-startup/src/banner/mod.rs
@@ -1,6 +1,6 @@
 use core::time::Duration;
 
-use themes::{BannerTheme, THEME_DEFAULT};
+use themes::BannerTheme;
 use vex_sdk::{
     vexBatteryCapacityGet, vexCompetitionStatus, vexSystemPowerupTimeGet, vexSystemVersion,
 };
@@ -9,9 +9,8 @@ use vexide_core::println;
 pub mod themes;
 
 #[inline]
-pub(crate) fn print() {
+pub(crate) fn print(theme: &BannerTheme) {
     const VEXIDE_VERSION: &str = "0.3.0";
-    const THEME: BannerTheme = THEME_DEFAULT;
 
     let system_version = unsafe { vexSystemVersion() }.to_be_bytes();
     let competition_status = unsafe { vexCompetitionStatus() };
@@ -29,17 +28,17 @@ pub(crate) fn print() {
 {lp6}          +%%%*:   .+###%#\x1B[0m     ├─\x1B{mk}🔋 Battery:\x1B[0m {battery}%
 {lp7}           .%:\x1B[0m                 ╰─\x1B{mk}⌚ Uptime:\x1B[0m {uptime:.2?}
 ",
-        lp1 = THEME.logo_primary[0],
-        lp2 = THEME.logo_primary[1],
-        lp3 = THEME.logo_primary[2],
-        lp4 = THEME.logo_primary[3],
-        lp5 = THEME.logo_primary[4],
-        lp6 = THEME.logo_primary[5],
-        lp7 = THEME.logo_primary[6],
-        ls = THEME.logo_secondary,
-        cv = THEME.crate_version,
-        mk = THEME.metadata_key,
-        emoji = THEME.emoji,
+        lp1 = theme.logo_primary[0],
+        lp2 = theme.logo_primary[1],
+        lp3 = theme.logo_primary[2],
+        lp4 = theme.logo_primary[3],
+        lp5 = theme.logo_primary[4],
+        lp6 = theme.logo_primary[5],
+        lp7 = theme.logo_primary[6],
+        ls = theme.logo_secondary,
+        cv = theme.crate_version,
+        mk = theme.metadata_key,
+        emoji = theme.emoji,
         vexide_version = VEXIDE_VERSION,
         vexos_version = format_args!(
             "{}.{}.{}-r{}",